code/requirements.txt
torch>=2.1
tqdm>=4.66
numpy>=1.24
Pillow>=10.0
torch-fidelity>=0.3.0
diffusers>=0.30
